legacy/requirements.txt
requests
httpx
python-whois
colorama
python-dotenv
groq
google-genai
fpdf
